Revert "Prevent unsafe narrowing: base/strings"
This reverts commit c2df0c99dff9e53302ab347266628229dffccd5b.

Reason for revert: cause test failure on win7 https://ci.chromium.org/ui/p/chromium/builders/ci/Win7%20Tests%20(1)/129448/test-results

Original change's description:
> Prevent unsafe narrowing: base/strings
>
> Bug: 1292951
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3696478
> Owners-Override: danakj <danakj@chromium.org>
> Commit-Queue: danakj <danakj@chromium.org>
> Commit-Queue: Peter Kasting <pkasting@chromium.org>
> Auto-Submit: Peter Kasting <pkasting@chromium.org>
> Cr-Commit-Position: refs/heads/main@{#1014622}

Bug: 1292951
No-Presubmit: true
No-Tree-Checks: true
No-Try: true
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3708881
Bot-Commit: Rubber Stamper <rubber-stamper@appspot.gserviceaccount.com>
Auto-Submit: Phillis Tang <phillis@chromium.org>
Owners-Override: Phillis Tang <phillis@chromium.org>
Commit-Queue: Rubber Stamper <rubber-stamper@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1014684}

+++ chromeos/network/shill_property_util.cc
@@ -32,7 +32,7 @@ namespace {
 // Replace non UTF8 characters in |str| with a replacement character.
 std::string ValidateUTF8(const std::string& str) {
   std::string result;
-  for (size_t index = 0; index < str.size(); ++index) {
+  for (int32_t index = 0; index < static_cast<int32_t>(str.size()); ++index) {
     base_icu::UChar32 code_point_out;
     bool is_unicode_char = base::ReadUnicodeCharacter(str.c_str(), str.size(),
                                                       &index, &code_point_out);
